# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zonas Colaborativas API
  version: 0.1.0
servers:
  - url: http://localhost:3001
paths:
...
  /boards/{boardId}/zones:
    get:
      summary: List zones
      parameters:
        - in: path
          name: boardId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Create zone
      parameters:
        - in: path
          name: boardId
          required: true
          schema: { type: string }
      responses:
        '201': { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...